+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2020-04-28
+### Added
+- static method to read OAuth Server config
+### Changed
+- check supported signing algorithms to determine whether to use token introspection endpoint
+
 ## [3.1.0] - 2020-04-27
 ### Changed
 - set auth header for token introspection calls instead of registering a client request filter